Make daft-clis python support optional
